+++ st2actions/st2actions/cmd/st2resultstracker.py
@@ -9,6 +9,7 @@ from st2common.models.db import db_setup
 from st2common.models.db import db_teardown
 from st2common.constants.logging import DEFAULT_LOGGING_CONF_PATH
 from st2common.transport.utils import register_exchanges
+from st2common.signal_handlers import register_common_signal_handlers
 from st2actions import config
 from st2actions import resultstracker
 
@@ -40,6 +41,7 @@ def _setup():
     db_setup(cfg.CONF.database.db_name, cfg.CONF.database.host, cfg.CONF.database.port,
              username=username, password=password)
     register_exchanges()
+    register_common_signal_handlers()
 
 
 def _run_worker():