+++ st2api/st2api/cmd/api.py
@@ -25,6 +25,7 @@ from st2common.models.db import db_setup
 from st2common.models.db import db_teardown
 from st2common.constants.logging import DEFAULT_LOGGING_CONF_PATH
 from st2common.transport.utils import register_exchanges
+from st2common.signal_handlers import register_common_signal_handlers
 from st2api.listener import get_listener_if_set
 from st2api import config
 from st2api import app
@@ -58,6 +59,7 @@ def _setup():
     db_setup(cfg.CONF.database.db_name, cfg.CONF.database.host, cfg.CONF.database.port,
              username=username, password=password)
     register_exchanges()
+    register_common_signal_handlers()
 
 
 def _run_server():